// src/lib/supabase/types.ts
export type Database = {
  public: {
    Tables: {
      bookings: {
        Row: {
          circuit_number: number
          created_at: string
          id: string
          number_of_pilots: number
          payment_status: string
          slot_id: string
          updated_at: string
          user_id: string
        }
        Insert: {
          circuit_number: number
          created_at?: string
          id?: string
          number_of_pilots: number
          payment_status?: string
          slot_id: string
          updated_at?: string
          user_id: string
        }
        Update: {
          circuit_number?: number
          created_at?: string
          id?: string
          number_of_pilots?: number
          payment_status?: string
          slot_id?: string
          updated_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "bookings_slot_id_fkey"
            columns: ["slot_id"]
            isOneToOne: false
            referencedRelation: "slots"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "bookings_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      message_status: "pending" | "replied"
      user_role: "admin" | "user"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}